refactor(Kafka Trigger Node): Name the v2 consumer modules in PascalCase (no-changelog)
nodes-base overwhelmingly uses PascalCase for multi-word files, 1146 against 11,
even though packages/nodes-base/AGENTS.md says kebab-case. Follow the code rather
than the doc, as raised in review.

The integration tests keep their current names: they are named after the
scenario rather than mirroring a module, and the documented way to run them
filters on that name.
