+++ packages/nodes-base/nodes/Kafka/v2/consumer/MessageParser.ts
