+++ packages/nodes-base/nodes/Kafka/v2/consumer/index.ts
@@ -2,15 +2,15 @@ export {
 	consumeTopic,
 	DEFAULT_BATCH_SIZE,
 	DEFAULT_PARTITIONS_CONSUMED_CONCURRENTLY,
-} from './consume-topic';
-export type { ConsumeTopicOptions, KafkaConsumerHandle } from './consume-topic';
-export { createDataEmitter } from './data-emitter';
+} from './ConsumeTopic';
+export type { ConsumeTopicOptions, KafkaConsumerHandle } from './ConsumeTopic';
+export { createDataEmitter } from './DataEmitter';
 export type {
 	DataEmitter,
 	DataEmitterContext,
 	DataEmitterOptions,
 	EmitResult,
 	ResolveOffsetMode,
-} from './data-emitter';
-export { createMessageParser } from './message-parser';
-export type { KafkaMessageParser, KafkaMessageParserOptions } from './message-parser';
+} from './DataEmitter';
+export { createMessageParser } from './MessageParser';
+export type { KafkaMessageParser, KafkaMessageParserOptions } from './MessageParser';
